[ADD] Improved builder and added new templates
A watcher option was added to auto build on file changes. It detects the
changes in the js scripts, and the css, html, and xml files inside the
src directory.

The css was updated to better represent deploy with its new colors and
logo.

A new way to use blocks was added to reuse code across different
templates. Also, a config json can be added at the top of the templates
to change the resulting fields in the xml.